S03P31D204-233 [+] Add update recipe rate

+++ backend/Frecipe/src/main/java/com/boum/frecipe/domain/recipe/Recipe.java
@@ -37,14 +37,14 @@ public class Recipe {
 	private Long view;
 	
 	// 평점
-	private float rate;
+	private double rate;
 		
 	public void updateContent(String title, String content) {
 		this.title = title;
 		this.content = content;
 	}
 	
-	public void updateRate(float rate) {
+	public void updateRate(double rate) {
 		this.rate = rate;
 	}
 }
